Support Array extension.
Array reference were always converted to classic java array. That produce invalid java code when a class extends Array directly. This CL fixes this problem by creating a reference to the Array type in that case.

Unfortunately, we cannot add a test for that case because j2cl generated build test creates a dependency to elemental2.

PiperOrigin-RevId: 135945415