Replace unsafe asArray call with asList.
When asArray is used as:

  SomeObject[] str=arrayLike.asArray()

it was resulting in an erasure cast to be added at call-site which could have failed unless the object is an array.

We cannot disable the cast checking for asArray via @UncheckedCast as it would result in unsafe code.

The patch replaces the API with asList which is still for-each loop friendly and doesn't have the same type safety issue.

PiperOrigin-RevId: 163712576